refactor(forge): split GitHub access methods into github_access.go (#1025)
Pure same-package code motion: move ProjectRole, roleForGitHubPermissions,
and DefaultBranchProtection out of github.go into a new sibling
github_access.go, the GitHub access/role/branch-protection seam. No
behaviour change; go doc -u -all is byte-identical.
